requirements.txt
requests~=2.28.1
lxml~=4.9.1
python-dateutil~=2.8.2